added PREVENT_STUCK_MODIFIER to bface_iso keymap

+++ tmk_core/protocol/lufa/descriptor.c
@@ -297,7 +297,7 @@ const USB_Descriptor_Configuration_t PROGMEM ConfigurationDescriptor =
 
             .ConfigAttributes       = (USB_CONFIG_ATTR_RESERVED | USB_CONFIG_ATTR_REMOTEWAKEUP),
 
-            .MaxPowerConsumption    = (uint8_t)USB_CONFIG_POWER_MA(USB_MAX_POWER_CONSUMPTION)
+            .MaxPowerConsumption    = USB_CONFIG_POWER_MA(USB_MAX_POWER_CONSUMPTION)
         },
 
     /*
